+++ common/config.ts
@@ -15,7 +15,7 @@ import { parseExpression } from "$common/expression_parser.ts";
 import type { System } from "$lib/plugos/system.ts";
 import type { ConfigObject } from "../plugs/index/config.ts";
 import { deepObjectMerge } from "@silverbulletmd/silverbullet/lib/json";
-import type { ActionButton } from "$lib/web.ts";
+import type { ActionButton } from "@silverbulletmd/silverbullet/type/client";
 
 const yamlConfigRegex = /^(```+|~~~+)(ya?ml|space-config)\r?\n([\S\s]+?)\1/m;
