+++ plugs/emoji/emoji.ts
@@ -1,7 +1,7 @@
 import emojiBlob from "./emoji.json" with { type: "json" };
 import type { CompleteEvent } from "../../plug-api/types.ts";
 import { editor, system } from "@silverbulletmd/silverbullet/syscalls";
-import type { EmojiConfig } from "$lib/web.ts";
+import type { EmojiConfig } from "@silverbulletmd/silverbullet/type/client";
 
 let emojiConfig: EmojiConfig = { aliases: [] };
