+++ plugs/plug-manager/plugmanager.ts
@@ -6,7 +6,6 @@ import {
 } from "@silverbulletmd/silverbullet/syscalls";
 import { readYamlPage } from "@silverbulletmd/silverbullet/lib/yaml_page";
 import { builtinPlugNames } from "../builtin_plugs.ts";
-import { plugPrefix } from "$common/spaces/constants.ts";
 
 const plugsPrelude =
   "This file lists all plugs that SilverBullet will load. Run the {[Plugs: Update]} command to update and reload this list of plugs.\n\n";
@@ -78,7 +77,7 @@ export async function updatePlugsCommand() {
       allCustomPlugNames.push(plugName);
       // console.log("Writing", `_plug/${plugName}.plug.js`, workerCode);
       await space.writeAttachment(
-        `${plugPrefix}${plugName}.plug.js`,
+        `_plug/${plugName}.plug.js`,
         new TextEncoder().encode(workerCode),
       );
     }
@@ -87,7 +86,7 @@ export async function updatePlugsCommand() {
     // And delete extra ones
     for (const { name: existingPlug } of await space.listPlugs()) {
       const plugName = existingPlug.substring(
-        plugPrefix.length,
+        "_plug/".length,
         existingPlug.length - ".plug.js".length,
       );
       if (!allPlugNames.includes(plugName)) {
