+++ web/components/panel.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from "preact/hooks";
 import type { Client } from "../client.ts";
-import type { PanelConfig } from "../../type/web.ts";
+import type { PanelConfig } from "../type.ts";
 import { panelHtml } from "./panel_html.ts";
 
 export function Panel({
