+++ web/components/top_bar.tsx
@@ -3,7 +3,7 @@ import type {
   CompletionResult,
 } from "@codemirror/autocomplete";
 import type { ComponentChildren, FunctionalComponent } from "preact";
-import type { Notification } from "$lib/web.ts";
+import type { Notification } from "@silverbulletmd/silverbullet/type/client";
 import type { FeatherProps } from "preact-feather/types";
 import type { IconBaseProps } from "react-icons/types";
 import { MiniEditor } from "./mini_editor.tsx";
