+++ web/reducer.ts
@@ -1,5 +1,5 @@
 import type { PageMeta } from "../plug-api/types.ts";
-import type { Action, AppViewState } from "../type/web.ts";
+import type { Action, AppViewState } from "./type.ts";
 
 export default function reducer(
   state: AppViewState,
